Replace API route files with api.js
I deleted all files in routes/api/ and replaced them all with one simple api.js file.
It contains all routing for the API, as well as all the route validation checks.
It didn't seem necessary to separate all this logic, things were too spread out
and hard to find.

The idea of having routes for a public and internal API also seemed unnecessary,
so those related files like /public/index.js were deleted.

The root index.js got an update in order to find the new routes file, api.js.
