+++ index.js
@@ -7,7 +7,7 @@ const db = require("./util/db");
 const errorHandler = require("./middleware/error");
 const log = require("./middleware/log");
 const isLoggedIn = require("./middleware/isLoggedIn");
-const apiRoutes = require("./routes/api/index");
+const apiRoutes = require("./routes/api");
 const pagesRoutes = require("./routes/pages");
 const authRoutes = require("./routes/auth");
 const app = express();
